requirements/base.txt
amqp==5.2.0
asgiref==3.8.1
attrs==22.2.0
billiard==4.2.0
build==0.10.0
CacheControl==0.12.11
celery==5.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
cron-descriptor==1.4.3
cryptography==39.0.2
distlib==0.3.6
Django==4.2.10
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-filter==24.2
django-jalali-date==1.1.2
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
dulwich==0.21.3
filelock==3.9.0
gunicorn==22.0.0
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
installer==0.6.0
jalali_core==1.0.0
jaraco.classes==3.2.3
jdatetime==5.0.0
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
kombu==5.3.7
lockfile==0.12.2
Markdown==3.6
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.0
pexpect==4.8.0
pillow==10.3.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
prompt_toolkit==3.0.45
psycopg==3.1.19
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==2.15.1
redis==5.0.4
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.5.0
tomlkit==0.11.6
trove-classifiers==2023.3.9
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.15
vine==5.1.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
wcwidth==0.2.13
webencodings==0.5.1
zipp==3.15.0
